搭建首页框架完成

+++ LYQdyzb/LYQdyzb/Classes/Tools/Extension/Common.swift
@@ -8,8 +8,8 @@
 
 import UIKit
 
-let kStatusBarH = 20 //状态栏高度
-let kNavigationBarH = 44 //导航栏高度
+let kStatusBarH = (CGFloat)(20) //状态栏高度
+let kNavigationBarH = (CGFloat)(44) //导航栏高度
 
 let kScreenW = UIScreen.main.bounds.width //屏幕宽
 let kScreenH = UIScreen.main.bounds.height //屏幕高
